float to BigDecimal and initial market making

+++ backend/src/main/java/com/bitcorner/service/BalanceService.java
@@ -4,14 +4,15 @@ import com.bitcorner.entity.Balance;
 
 import javax.management.BadAttributeValueExpException;
 import javax.persistence.EntityNotFoundException;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface BalanceService {
     void save(Balance balance);
     List<Balance> getAllBalance(String userId);
     Balance getBalance(String userId, long currencyId);
-    void depositBalance(String userId,long currencyId,float depositAmount) throws BadAttributeValueExpException;
+    void depositBalance(String userId, long currencyId, BigDecimal depositAmount) throws BadAttributeValueExpException;
 
-    void withdrawBalance(String userId,long currencyId,float withdrawAmount) throws BadAttributeValueExpException;
+    void withdrawBalance(String userId,long currencyId,BigDecimal withdrawAmount) throws BadAttributeValueExpException;
 
 }
